+++ js/gutenberg-voucher-sales.js
@@ -67,7 +67,7 @@ registerGutenbergBlock('recras/voucher-sales', {
                     redirect: newVal
                 });
             },
-            options: pagesPosts,
+            children: pagesPosts,
             placeholder: wp.i18n.__('i.e. https://www.recras.com/thanks/', TEXT_DOMAIN),
             label: wp.i18n.__('Thank-you page (optional, leave empty to not redirect)', TEXT_DOMAIN),
         };