+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Thank-you page list is now better split into pages and posts
+
 ## 5.3.1 (2023-03-14)
 * Fix GA4 events not being sent on book process thank-you page